feat: schdule 기능

+++ services/schedule.service.js
@@ -1,21 +1,44 @@
-import { ScheduleRepository } from '../repositories';
+import { ScheduleRepository, UserRepository, TrainerRepository } from '../repositories';
 class ScheduleService {
   _scheduleRepository = new ScheduleRepository();
+  _userRepository = new UserRepository();
+  _trainerRepository = new TrainerRepository();
 
-  postSchedule = async (title, date) => {
+  // 스케줄 생성
+  postSchedule = async (title, date, userId) => {
     try {
       if (!title) {
         return {
           status: 400,
           message: '제목 미입력',
         };
-      } else if (!date) {
+      }
+
+      if (!date) {
         return {
           status: 400,
           message: '날짜 미입력',
         };
       }
-      const result = await this._scheduleRepository.postSchedule(title, date);
+      const user = await this._userRepository.getUserById(userId);
+
+      if (user.type !== 'trainer' || user.type !== 'admin') {
+        return {
+          status: 400,
+          message: '스케줄 생성 권한 없음',
+        };
+      }
+
+      const trainer = await this._trainerRepository.getTrainerInfo(userId);
+
+      if (!trainer) {
+        return {
+          status: 400,
+          message: '트레이너 정보 없음',
+        };
+      }
+
+      const result = await this._scheduleRepository.postSchedule(title, date, userId, trainer.trainerId);
       if (!result) {
         return {
           status: 400,
@@ -31,7 +54,22 @@ class ScheduleService {
     }
   };
 
-  oneGetSchedule = async scheduleId => {
+  // 스케줄 전체 조회
+  allGetSchedule = async userId => {
+    try {
+      const result = await this._scheduleRepository.allGetSchedule(userId);
+
+      return {
+        status: 200,
+        message: result,
+      };
+    } catch (err) {
+      return { status: 500, message: 'Server Error' };
+    }
+  };
+
+  // 스케줄 상세 조회
+  oneGetSchedule = async (scheduleId, userId) => {
     try {
       if (!scheduleId) {
         return {
@@ -39,6 +77,16 @@ class ScheduleService {
           message: '스케줄 ID 미입력',
         };
       }
+
+      const user = await this._userRepository.getUserById(userId);
+
+      if (user.type !== 'trainer' || user.type !== 'admin') {
+        return {
+          status: 400,
+          message: '스케줄 조회 권한 없음',
+        };
+      }
+
       const result = await this._scheduleRepository.oneGetSchedule(scheduleId);
       if (!result) {
         return {
@@ -55,7 +103,8 @@ class ScheduleService {
     }
   };
 
-  putSchedule = async (scheduleId, title, date) => {
+  // 스케줄 수정
+  putSchedule = async (scheduleId, title, date, userId) => {
     try {
       if (!scheduleId) {
         return {
@@ -63,23 +112,38 @@ class ScheduleService {
           message: '스케줄 ID 미입력',
         };
       }
+
       const schedule = await this._scheduleRepository.oneGetSchedule(scheduleId);
       if (!schedule) {
         return {
           status: 400,
           message: '존재하지 않는 스케줄 ID',
         };
-      } else if (!title) {
+      }
+
+      const user = await this._userRepository.getUserById(userId);
+
+      if (schedule.userId !== user.userId || user.type !== 'admin') {
+        return {
+          status: 400,
+          message: '스케줄 수정 권한 없음',
+        };
+      }
+
+      if (!title) {
         return {
           status: 400,
           message: '제목 미입력',
         };
-      } else if (!date) {
+      }
+
+      if (!date) {
         return {
           status: 400,
           message: '날짜 미입력',
         };
       }
+
       const result = await this._scheduleRepository.putSchedule(scheduleId, title, date);
       if (!result) {
         return {
@@ -95,7 +159,9 @@ class ScheduleService {
       return { status: 500, message: 'Server Error' };
     }
   };
-  deleteSchedule = async scheduleId => {
+
+  // 스케줄 삭제
+  deleteSchedule = async (scheduleId, userId) => {
     try {
       if (!scheduleId) {
         return {
@@ -103,6 +169,7 @@ class ScheduleService {
           message: '스케줄 ID 미입력',
         };
       }
+
       const schedule = await this._scheduleRepository.oneGetSchedule(scheduleId);
       if (!schedule) {
         return {
@@ -110,6 +177,16 @@ class ScheduleService {
           message: '존재하지 않는 스케줄 ID',
         };
       }
+
+      const user = await this._userRepository.getUserById(userId);
+
+      if (schedule.userId !== user.userId || user.type !== 'admin') {
+        return {
+          status: 400,
+          message: '스케줄 삭제 권한 없음',
+        };
+      }
+
       const result = await this._scheduleRepository.deleteSchedule(scheduleId);
       if (!result) {
         return {
